+++ modelsTimescale/detections/detection.js
@@ -1,11 +1,5 @@
 module.exports = (sequelize, DataTypes) => {
   const Detection = sequelize.define('Detection', {
-    id: {
-      type: DataTypes.UUID,
-      defaultValue: DataTypes.UUIDV4,
-      allowNull: false,
-      primaryKey: true
-    },
     start: {
       // Hypertable key
       type: DataTypes.DATE(3)
@@ -28,15 +22,15 @@ module.exports = (sequelize, DataTypes) => {
   }, {
     timestamps: false
   })
+  Detection.removeAttribute('id') // https://github.com/sequelize/sequelize/issues/1026#issuecomment-54877327
   Detection.associate = function (models) {
     Detection.belongsTo(models.Stream, { as: 'stream', foreignKey: 'stream_id' })
     Detection.belongsTo(models.Classification, { as: 'classification', foreignKey: 'classification_id' })
     Detection.belongsTo(models.Classifier, { as: 'classifier', foreignKey: 'classifier_id' })
-    Detection.hasMany(models.DetectionReview, { as: 'reviews', foreignKey: 'detection_id' })
   }
   Detection.attributes = {
-    lite: ['id', 'stream_id', 'start', 'end', 'confidence'],
-    full: ['id', 'stream_id', 'start', 'end', 'confidence']
+    lite: ['stream_id', 'start', 'end', 'confidence'],
+    full: ['stream_id', 'start', 'end', 'confidence']
   }
   return Detection
 }