+++ routes/core/detections/stream.js
@@ -33,10 +33,6 @@ const auth0Service = require('../../../services/auth0/auth0-service')
  *         in: query
  *         required: true
  *         type: string
- *       - name: reviews
- *         description: Whether or not to include detection user reviews or not
- *         in: query
- *         type: boolean
  *       - name: classifications
  *         description: List of clasification identifiers
  *         in: query
@@ -57,17 +53,12 @@ const auth0Service = require('../../../services/auth0/auth0-service')
  *         default: 0
  *     responses:
  *       200:
- *         description: List of detections objects. **"reviews" attribute is included based on "reviews" query parameter**
  *         content:
  *           application/json:
  *             schema:
  *               type: array
  *               items:
- *                 oneOf:
- *                   - $ref: '#/components/schemas/Detection'
- *                   - $ref: '#/components/schemas/DetectionWithReviews'
- *                 discriminator:
- *                   propertyName: reviews
+ *                 $ref: '#/components/schemas/Detection'
  *       400:
  *         description: Invalid query parameters
  *       404:
@@ -84,7 +75,6 @@ router.get('/:id/detections', function (req, res) {
   params.convert('min_confidence').optional().toFloat()
   params.convert('limit').optional().toInt()
   params.convert('offset').optional().toInt()
-  params.convert('reviews').optional().toBoolean()
 
   return params.validate()
     .then(async () => {
@@ -98,10 +88,10 @@ router.get('/:id/detections', function (req, res) {
       }
     })
     .then(async () => {
-      const { start, end, classifications, limit, offset, reviews } = convertedParams
+      const { start, end, classifications, limit, offset } = convertedParams
       const minConfidence = convertedParams.min_confidence
-      const detections = await detectionsService.query(start, end, streamId, classifications, minConfidence, reviews, limit, offset, user)
-      return res.json(detections)
+      const result = await detectionsService.query(start, end, streamId, classifications, minConfidence, limit, offset, user)
+      return res.json(result)
     })
     .catch(httpErrorHandler(req, res, 'Failed getting detections'))
 })